Added BFCE loss to loss_functions.py
